+++ examples/data-sources/nxos_default_qos_class_map_dscp/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_default_qos_class_map_dscp" "example" {
+  class_map_name = "Voice"
+  value          = "ef"
+}