+++ examples/resources/nxos_default_qos_class_map_dscp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_default_qos_class_map_dscp.example "sys/ipqos/dflt/c/name-[Voice]/dscp-[ef]"