+++ examples/resources/nxos_default_qos_class_map_dscp/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_default_qos_class_map_dscp" "example" {
+  class_map_name = "Voice"
+  value          = "ef"
+}