+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Add nxos_pim_ssm_policy resource and data source
 - Add nxos_pim_ssm_range resource and data source
 - Add nxos_default_qos_class_map resource and data source
+- Add nxos_default_qos_class_map_dscp resource and data source
 
 ## 0.2.3